requirements.txt
opencv-python
numpy
requests
python-dotenv
flask
websockets